Flow log files now .csv, improve flow and utilization plot

+++ build/plot_helpers/flow_plot/flow_plot.py
@@ -1,23 +1,33 @@
 import sys
-from exputil import *
+import exputil
 
-try:
-    from .generate_rate_txt import generate_rate_txt
-except (ImportError, SystemError):
-    from generate_rate_txt import generate_rate_txt
+from generate_flow_rate_csv import generate_flow_rate_csv
 
 
-def flow_plot(logs_ns3_dir, data_out_dir, pdf_out_dir, flow_dir, interval_ns, flow_id):
-    local_shell = LocalShell()
+def flow_plot(logs_ns3_dir, data_out_dir, pdf_out_dir, flow_id, interval_ns):
+    local_shell = exputil.LocalShell()
+
+    # Check that all plotting files are available
+    if not local_shell.file_exists("plot_tcp_time_vs_cwnd.plt") or \
+       not local_shell.file_exists("plot_tcp_time_vs_progress.plt") or \
+       not local_shell.file_exists("plot_tcp_time_vs_rtt.plt") or \
+       not local_shell.file_exists("plot_tcp_time_vs_rate.plt"):
+        print("The gnuplot files are not present.")
+        print("Are you executing this python file inside the flow_plot directory?")
+        exit(1)
+
+    # Create the output directories if they don't exist yet
+    local_shell.make_full_dir(data_out_dir)
+    local_shell.make_full_dir(pdf_out_dir)
     
     # Create rate file
-    generate_rate_txt(logs_ns3_dir, data_out_dir, flow_id, interval_ns)
+    generate_flow_rate_csv(logs_ns3_dir, data_out_dir, flow_id, interval_ns)
     
     # Plot time vs. cwnd
-    data_filename = logs_ns3_dir + "/flow_" + str(flow_id) + "_cwnd.txt"
-    local_shell.copy_file(data_filename, data_out_dir + "/flow_" + str(flow_id) + "_cwnd.txt")
+    data_filename = logs_ns3_dir + "/flow_" + str(flow_id) + "_cwnd.csv"
+    local_shell.copy_file(data_filename, data_out_dir + "/flow_" + str(flow_id) + "_cwnd.csv")
     pdf_filename = pdf_out_dir + "/plot_tcp_time_vs_cwnd_" + str(flow_id) + ".pdf"
-    plt_filename = flow_dir + "/plot_tcp_time_vs_cwnd.plt"
+    plt_filename = "plot_tcp_time_vs_cwnd.plt"
     local_shell.copy_file(plt_filename, "temp.plt")
     local_shell.sed_replace_in_file_plain("temp.plt", "[OUTPUT-FILE]", pdf_filename)
     local_shell.sed_replace_in_file_plain("temp.plt", "[DATA-FILE]", data_filename)
@@ -26,10 +36,10 @@ def flow_plot(logs_ns3_dir, data_out_dir, pdf_out_dir, flow_dir, interval_ns, fl
     local_shell.remove("temp.plt")
 
     # Plot time vs. progress
-    data_filename = logs_ns3_dir + "/flow_" + str(flow_id) + "_progress.txt"
-    local_shell.copy_file(data_filename, data_out_dir + "/flow_" + str(flow_id) + "_progress.txt")
+    data_filename = logs_ns3_dir + "/flow_" + str(flow_id) + "_progress.csv"
+    local_shell.copy_file(data_filename, data_out_dir + "/flow_" + str(flow_id) + "_progress.csv")
     pdf_filename = pdf_out_dir + "/plot_tcp_time_vs_progress_" + str(flow_id) + ".pdf"
-    plt_filename = flow_dir + "/plot_tcp_time_vs_progress.plt"
+    plt_filename = "plot_tcp_time_vs_progress.plt"
     local_shell.copy_file(plt_filename, "temp.plt")
     local_shell.sed_replace_in_file_plain("temp.plt", "[OUTPUT-FILE]", pdf_filename)
     local_shell.sed_replace_in_file_plain("temp.plt", "[DATA-FILE]", data_filename)
@@ -38,10 +48,10 @@ def flow_plot(logs_ns3_dir, data_out_dir, pdf_out_dir, flow_dir, interval_ns, fl
     local_shell.remove("temp.plt")
 
     # Plot time vs. rtt
-    data_filename = logs_ns3_dir + "/flow_" + str(flow_id) + "_rtt.txt"
-    local_shell.copy_file(data_filename, data_out_dir + "/flow_" + str(flow_id) + "_rtt.txt")
+    data_filename = logs_ns3_dir + "/flow_" + str(flow_id) + "_rtt.csv"
+    local_shell.copy_file(data_filename, data_out_dir + "/flow_" + str(flow_id) + "_rtt.csv")
     pdf_filename = pdf_out_dir + "/plot_tcp_time_vs_rtt_" + str(flow_id) + ".pdf"
-    plt_filename = flow_dir + "/plot_tcp_time_vs_rtt.plt"
+    plt_filename = "plot_tcp_time_vs_rtt.plt"
     local_shell.copy_file(plt_filename, "temp.plt")
     local_shell.sed_replace_in_file_plain("temp.plt", "[OUTPUT-FILE]", pdf_filename)
     local_shell.sed_replace_in_file_plain("temp.plt", "[DATA-FILE]", data_filename)
@@ -50,9 +60,9 @@ def flow_plot(logs_ns3_dir, data_out_dir, pdf_out_dir, flow_dir, interval_ns, fl
     local_shell.remove("temp.plt")
 
     # Plot time vs. rate
-    data_filename = data_out_dir + "/flow_" + str(flow_id) + "_rate.txt"
+    data_filename = data_out_dir + "/flow_" + str(flow_id) + "_rate.csv"
     pdf_filename = pdf_out_dir + "/plot_tcp_time_vs_rate_" + str(flow_id) + ".pdf"
-    plt_filename = flow_dir + "/plot_tcp_time_vs_rate.plt"
+    plt_filename = "plot_tcp_time_vs_rate.plt"
     local_shell.copy_file(plt_filename, "temp.plt")
     local_shell.sed_replace_in_file_plain("temp.plt", "[OUTPUT-FILE]", pdf_filename)
     local_shell.sed_replace_in_file_plain("temp.plt", "[DATA-FILE]", data_filename)
@@ -63,19 +73,17 @@ def flow_plot(logs_ns3_dir, data_out_dir, pdf_out_dir, flow_dir, interval_ns, fl
 
 def main():
     args = sys.argv[1:]
-    if len(args) != 6:
-        print("Must supply exactly six arguments")
-        print("Usage: python flow_plot.py [logs_ns3 directory] [data_out_dir] [pdf_out_dir] [flow_dir] "
-              "[interval_ns] [flow_id]")
+    if len(args) != 5:
+        print("Must supply exactly five arguments")
+        print("Usage: python flow_plot.py [logs_ns3_dir] [data_out_dir] [pdf_out_dir] [flow_id] [interval_ns]")
         exit(1)
     else:
         flow_plot(
             args[0],
             args[1],
             args[2],
-            args[3],
-            int(args[4]),
-            int(args[5])
+            int(args[3]),
+            int(args[4])
         )
